requirements.txt
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.3.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.1.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kaleido==0.0.3.post1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.1
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.19.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
portpicker==1.3.1
prometheus-client==0.8.0
prompt-toolkit==3.0.6
psutil==5.7.2
ptyprocess==0.6.0
pycparser==2.20
pydataset==0.2.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
svgpathtools==1.3.3
svgwrite==1.4
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toolz==0.10.0
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
vega==3.4.0
vega-datasets==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0